requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
python-dotenv
pypdf
chromadb 
langchain-chroma
beautifulsoup4 == 4.14.3